feat(mind-map): let a node wear a Trilium icon
Extracts the icon picker out of the note icon widget so that anything can be
given an icon through it, and offers it for a node in the panel.

Mind Elixir renders an icon as text, so a class arrives as the words
themselves: the class is put onto the span it landed in after every layout,
leaving the icon pack's own CSS to draw it. Text that is not a class — the
emoji a map made in another tool carries — is left alone.
