+++ apps/client/src/translations/en/translation.json
@@ -3531,6 +3531,9 @@
     "font-size-medium": "M",
     "font-size-large": "L",
     "font-size-extra-large": "XL",
+    "icon": "Icon",
+    "choose-icon": "Choose an icon",
+    "clear-icon": "Remove the icon",
     "tags": "Tags",
     "tags-placeholder": "Type a tag and press Enter",
     "tags-differ": "The selected nodes don't carry the same tags, so these can only be read.",
