+++ apps/client/src/widgets/note_icon.css
@@ -39,55 +39,6 @@ body > div.note-icon-widget > .dropdown-menu {
     padding: 10px 15px 10px 15px !important;
 }
 
-.note-icon-widget .filter-row {
-    padding: 10px;
-    display: flex;
-    align-items: center;
-    gap: 1em;
-}
-
-.note-icon-widget .filter-row span {
-    display: block;
-    font-weight: bold;
-}
-
-.note-icon-widget .icon-list {
-    height: 500px;
-    overflow: auto;
-}
-
-.note-icon-widget .icon-list span {
-    display: inline-block;
-    padding: 10px;
-    cursor: pointer;
-    border: 1px solid transparent;
-    font-size: 180%;
-}
-
-.note-icon-widget .icon-list span:hover {
-    border: 1px solid var(--main-border-color);
-}
-
-.note-icon-widget .icon-list span.bx-empty {
-    width: unset;
-}
-
-.note-icon-widget .icon-list span.bx-empty::before {
-    display: inline-block;
-    content: "";
-    border: 1px dashed var(--muted-text-color);
-    width: 1em;
-    height: 1em;
-}
-
-.note-icon-widget {
-    .no-results {
-        padding: 20px;
-        text-align: center;
-        color: var(--muted-text-color);
-    }
-}
-
 body.experimental-feature-new-layout {
     .note-icon-widget button.note-icon {
         --input-focus-outline-color: var(--note-icon-hover-background-color);
